build: Move SSRF protection into @n8n/backend-network package
