+++ packages/@n8n/backend-network/src/dns/index.ts
@@ -1,0 +1,1 @@
+export { DnsResolver } from './dns-resolver';
