+++ packages/@n8n/backend-network/src/index.ts
@@ -1,1 +1,2 @@
-export {};
+export * from './dns';
+export * from './ssrf';
