+++ packages/@n8n/backend-network/src/ssrf/index.ts
@@ -1,0 +1,3 @@
+export { SsrfProtectionService } from './ssrf-protection.service';
+export type { SsrfBridge } from './ssrf-protection.service';
+export { SsrfBlockedIpError } from './ssrf-blocked-ip.error';
