+++ packages/@n8n/backend-network/tsconfig.json
@@ -7,5 +7,11 @@
 		"emitDecoratorMetadata": true
 	},
 	"include": ["src/**/*.ts"],
-	"references": [{ "path": "../backend-common/tsconfig.build.json" }]
+	"references": [
+		{ "path": "../../workflow/tsconfig.build.cjs.json" },
+		{ "path": "../backend-common/tsconfig.build.json" },
+		{ "path": "../config/tsconfig.build.json" },
+		{ "path": "../constants/tsconfig.build.json" },
+		{ "path": "../di/tsconfig.build.json" }
+	]
 }
