+++ packages/cli/src/modules/instance-ai/web-research/__tests__/fetch-and-extract.test.ts
@@ -1,12 +1,10 @@
 import type { Logger } from '@n8n/backend-common';
+import type { DnsResolver, SsrfBridge } from '@n8n/backend-network';
+import { SsrfProtectionService } from '@n8n/backend-network';
 import { SsrfProtectionConfig } from '@n8n/config';
 import { mock } from 'jest-mock-extended';
-import type { SsrfBridge } from 'n8n-core';
 import { createResultOk } from 'n8n-workflow';
 import type { LookupFunction } from 'node:net';
-
-import type { DnsResolver } from 'n8n-core';
-import { SsrfProtectionService } from 'n8n-core';
 
 import { fetchAndExtract } from '../fetch-and-extract';
