+++ packages/cli/src/services/__tests__/ai-workflow-builder.service.test.ts
@@ -14,7 +14,7 @@ import { LoadNodesAndCredentials } from '@/load-nodes-and-credentials';
 import type { Push } from '@/push';
 import { WorkflowBuilderService } from '@/services/ai-workflow-builder.service';
 import type { DynamicNodeParametersService } from '@/services/dynamic-node-parameters.service';
-import type { SsrfProtectionService } from 'n8n-core';
+import type { SsrfProtectionService } from '@n8n/backend-network';
 import type { UrlService } from '@/services/url.service';
 import type { Telemetry } from '@/telemetry';
 import type { WorkflowBuilderSessionRepository } from '@/modules/workflow-builder';
