+++ packages/core/package.json
@@ -55,6 +55,7 @@
     "@aws-sdk/client-s3": "3.808.0",
     "@langchain/core": "catalog:",
     "@n8n/backend-common": "workspace:*",
+    "@n8n/backend-network": "workspace:*",
     "@n8n/client-oauth2": "workspace:*",
     "@n8n/config": "workspace:*",
     "@n8n/constants": "workspace:*",
