+++ packages/core/src/execution-engine/index.ts
@@ -1,3 +1,4 @@
+import type { SsrfBridge } from '@n8n/backend-network';
 import type {
 	DataTableProxyProvider,
 	DynamicCredentialCheckProxyProvider,
@@ -7,8 +8,6 @@ import type {
 	IWorkflowSettings,
 	OauthJweProxyProvider,
 } from 'n8n-workflow';
-
-import type { SsrfBridge } from '@/ssrf';
 
 import type { ExecutionLifecycleHooks } from './execution-lifecycle-hooks';
 import type { ExternalSecretsProxy } from './external-secrets-proxy';
@@ -85,24 +84,24 @@ declare module 'n8n-workflow' {
 }
 
 export * from './active-workflow-triggers';
-export * from './scheduled-task-manager';
-export type * from './interfaces';
-export * from './routing-node';
-export * from './node-execution-context';
-export * from './partial-execution-utils';
-export * from './node-execution-context/utils/execution-metadata';
-export * from './workflow-execute';
-export * from './execution-context-hook-registry.service';
-export { ExecutionLifecycleHooks } from './execution-lifecycle-hooks';
-export { ExternalSecretsProxy, type IExternalSecretsManager } from './external-secrets-proxy';
-export { ExecutionContextService } from './execution-context.service';
-export { establishExecutionContext } from './execution-context';
-export { isEngineRequest } from './requests-response';
 export {
 	synthesizeBinaryFixture,
 	type FixtureSizeHint,
 	type SynthesizeBinaryFixtureOptions,
 } from './eval-mock-fixtures';
+export { establishExecutionContext } from './execution-context';
+export * from './execution-context-hook-registry.service';
+export { ExecutionContextService } from './execution-context.service';
+export { ExecutionLifecycleHooks } from './execution-lifecycle-hooks';
+export { ExternalSecretsProxy, type IExternalSecretsManager } from './external-secrets-proxy';
+export type * from './interfaces';
+export * from './node-execution-context';
+export * from './node-execution-context/utils/execution-metadata';
+export * from './partial-execution-utils';
+export { isEngineRequest } from './requests-response';
+export * from './routing-node';
+export * from './scheduled-task-manager';
+export * from './workflow-execute';
 // Exposed so eval-mode credential helpers (e.g. `EvalMockedCredentialsHelper`)
 // can reuse the same schema-driven cred synthesizer the wire-server URL
 // rewrite expects. See its `getDecrypted` catch path for the consumer.
