+++ packages/core/src/execution-engine/node-execution-context/utils/request-helpers/__tests__/legacy-request-adapter.test.ts
@@ -1,3 +1,4 @@
+import type { SsrfBridge } from '@n8n/backend-network';
 import FormData from 'form-data';
 import type { Agent as HttpsAgent } from 'https';
 import type {
@@ -12,7 +13,6 @@ import type { SecureContextOptions } from 'tls';
 import { mock } from 'vitest-mock-extended';
 
 import type { ExecutionLifecycleHooks } from '@/execution-engine/execution-lifecycle-hooks';
-import type { SsrfBridge } from '@/ssrf';
 
 import { parseRequestObject, proxyRequestToAxios } from '../legacy-request-adapter';
