+++ packages/core/src/index.ts
@@ -8,7 +8,6 @@ export * from './data-deduplication-service';
 export * from './encryption';
 export * from './errors';
 export * from './execution-engine';
-export * from './ssrf';
 export * from './html-sandbox';
 export * from './instance-settings';
 export * from './nodes-loader';
