--- packages/core/src/ssrf/index.ts
@@ -1,5 +1,0 @@
-export { DnsResolver } from './dns-resolver';
-export type { DnsLookupOptions } from './dns-resolver';
-export { SsrfBlockedIpError } from './ssrf-blocked-ip.error';
-export { SsrfProtectionService } from './ssrf-protection.service';
-export type { SsrfBridge, SsrfCheckResult, SsrfEventMap } from './ssrf-protection.service';
